Initiative 'filters in AngularJS'
Began the idea of creating AngularJS filters for reasons of those who
want to use the same masks used on inputs, also use in list, grid, etc.
I did the best of the code, but I tried to leave the most simple, clean
and not break the entire structure of you. Hope you like the idea and
strengthened the idea for everyone to use.