fix: return null for empty number and percent fields
Return null if the input of an ui-number-mask or
ui-percentage-mask field is empty instead of
returning undefined or an empty string.
This has the advantage, that no additional check is required for
null values inside of calculations (e.g. 5 + null === 5).
null is also a valid number, where as an empty string is not.
This is the same behavior as angular uses for input[number].