fix(ui-number-mask): show "0" if model is 0
Show 0, instead of undefined, if the model value has changed to 0.
When ui-number-mask="0" is used and the model changes the value
to 0, then undefined is displayed in the input field.
The cause of the error is, that clearDelimitersAndLeadingZero
does not only clear leading zeros, but also a single 0
(what is equal to the vaue 0). -> Do not strip a single 0.

Closes ##111