fix(uiPercentageMask): honor ui-hide-space when view value change to model update from controller (#260)
* fix(uiPercentageMask): honor ui-hide-space when view value change due to model update from controller

Previously, a change to the model in the controller would result in a space in between the percent sign and the value no matter if the ui-hide-space attribute is present or not.  Now, a change to the model will result in the correct view value format.

Non-breaking change.

* fix(uiPercentageMask): honor ui-hide-space when view value change due to model update from controller

Add missing mock inject in the new test.

Non-breaking change.

* fix(uiPercentageMask): honor ui-hide-space when view value change due to model update from controller

Update failing tests.  When accessing via rootScope, update model to 50 not 0.5 to represent 50%.

Non-breaking change.